Fix assembler code loading GDT.

+++ os/arch/i386/gdt.h
@@ -49,7 +49,7 @@ struct gdt_descriptor {
 #define GDT_FLAG_32_BIT 0x4
 #define GDT_FLAG_4_KiB_LIMIT 0x8
 
-void gdt_set_descriptor(uint32_t i, uint32_t base, uint32_t limit, uint8_t access, uint8_t flags);
+void gdt_set_descriptor(uint16_t i, uint32_t base, uint32_t limit, uint8_t access, uint8_t flags);
 void gdt_initialize();
 void gdt_install();
